moved state service for create form to the views-create-joke folder
